server/requirements.txt
absl-py==1.4.0
appdirs==1.4.4
arrow==1.2.3
asgiref==3.6.0
astunparse==1.6.3
brotlipy==0.7.0
cachetools==5.3.0
catalogue==2.0.8
certifi==2023.5.7
channels==4.0.0
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
contourpy==1.0.7
cookiecutter==2.1.1
cryptography==40.0.2
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
cymem==2.0.7
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.2
django-appconf==1.0.5
django-browser-reload==1.8.0
django-compressor==4.3.1
django-cors-headers==3.14.0
django-libsass==0.9
django-rest-framework==0.1.0
django-tailwind==3.5.0
djangorestframework==3.14.0
filelock==3.9.0
Flask==2.2.3
flatbuffers==23.1.21
fonttools==4.38.0
frozendict==2.3.8
grpcio==1.51.1
gunicorn==20.1.0
h5py==3.8.0
html5lib==1.1
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.2.0
jsonpointer==2.1
kiwisolver==1.4.4
langcodes==3.3.0
libclang==15.0.6.1
libsass==0.22.0
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.0
mpmath==1.2.1
multitasking==0.0.11
murmurhash==1.0.9
networkx==3.0
nltk==3.8.1
opencv-python==4.7.0.72
pandas==2.0.2
pathy==0.10.1
pdfminer==20191125
pdfminer.six==20221105
Pillow==9.5.0
premailer==3.10.0
preshed==3.0.8
prettyprint==0.1.5
protobuf==3.20.3
pycryptodome==3.18.0
pydantic==1.10.7
PyGithub==1.58.1
python-slugify==8.0.1
python-version==0.0.2
python3-openid==3.2.0
rcssmin==1.1.1
regex==2023.3.23
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
simplejson==3.19.1
smart-open==6.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spicy==0.16.0
sqlparse==0.4.4
srsly==2.4.6
sympy==1.11.1
text-unidecode==1.3
thinc==8.1.9
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==1.26.15
wasabi==1.1.1
webencodings==0.5.1
whitenoise==6.4.0
wincertstore==0.2
wrapt==1.15.0
yagmail==0.15.293
yahoo-finance==1.4.0
zstandard==0.18.0
